// src/pages/main-menu/payment/models/interface.ts
export interface CheckForm {
  bankId?: number;
  issueDateDate?: string;
  expiryDateDate?: string;
  checkNo?: string;
  issueDateText?: string;
  expiryDateText?: string;
}

export interface CreditOrDebitForm {
  cardNo?: string;
  issueDate?: string;
  expiryDate?: string;
}

export interface DeductDepositForm {
  issueDate?: string;
  expiryDate?: string;
}

export interface PaymentCreditPayload {
  fromWebFlag?: string;
  submitAmount?: string;
  returnAmount?: number;
  charge?: number;
  acctId?: number;
  partyType?: string;
  partyCode?: number;
  contactChannelId?: number;
  paymentMethodId?: number;
  acctResId?: number;
  prefix?: string;
  accNbr?: string;
  msisdn?: string;
  subsId?: number;
  acctItemMergeList?: any[];
  overdueQryFeeResultList?: any[];
  selectedAcctItemIdList?: any[];
  overdue?: number;
  remarks?: string;
  paymentSource?: string;
  isNeedCdr?: boolean;
  isSendAdvice?: boolean;
  isAllowRechargeBlack?: boolean;
  spId?: number;
  scratchCardPin?: string | null;
  check?: CheckForm;
  bankId?: string | null;
  checkNbr?: string | null;
  checkIssueDate?: string | null;
  checkExpDate?: string | null;
}

// Generated by https://quicktype.io

export interface ReceiptReq {
  rechargeAmount: string;
  rechargeTime: Date;
  acctNbr: string;
  balance: string;
  paymentId: string;
  paymentReason: string;
  preStoreAmount: string;
  settleAmount: string;
  accountType: string;
  partyType: string;
  partyCode: string;
  custName: string;
  receiptNo: string;
  acctResName: string;
  exchangeRate: string;
  currency: string;
  bankName: string;
  checkNbr: string;
  description: string;
  receiptDate: string;
  acctId: number;
  invoiceMonth: string;
  preBalance: string;
  subsId: number;
  serviceNumber: string;
  accNbrTypeName: string;
  creditLimit: number;
  lastBill: string;
  lastSecondBill: string;
  lastThirdBill: string;
  lastFourthBill: string;
  lastFifthBill: string;
  lastSixBill: string;
  aboveSixBill: string;
  totalPaid: string;
}